# Prepare Varnish-Cache for the age of AI