--- a/bin/varnishd/cache/cache_filter.h
+++ b/bin/varnishd/cache/cache_filter.h
@@ -318,6 +318,11 @@ vdpio_consolidate_vscarab(const struct vdp_ctx *vdc, struct vscarab *scarab)
 		scarab->used = f - scarab->s;
 }
 
+// Lifecycle management in cache_deliver_proc.c
+int VDPIO_Init(struct vdp_ctx *vdc, struct objcore *oc, struct ws *ws,
+    vai_notify_cb *notify_cb, void *notify_priv, struct vscaret *scaret);
+void VDPIO_Return(const struct vdp_ctx *vdc);
+void VDPIO_Fini(struct vdp_ctx *vdc);
 
 void v_deprecated_ VRT_AddVDP(VRT_CTX, const struct vdp *);
 void v_deprecated_ VRT_RemoveVDP(VRT_CTX, const struct vdp *);