… bin/varnishd/cache/cache_obj.c
… bin/varnishd/cache/cache_obj.h
… bin/varnishd/cache/cache_range.c
… bin/varnishd/http1/cache_http1_line.c
… bin/varnishd/storage/storage_persistent.c
… bin/varnishd/storage/storage_simple.c
… bin/varnishtest/tests/c00111.vtc
… bin/varnishtest/tests/m00061.vtc
… vmod/automake_boilerplate_debug.am
… vmod/vmod_debug.c
… vmod/vmod_debug.h
… vmod/vmod_debug.vcc
… vmod/vmod_debug_transport_vai.c