@nigoroll 2024-10-14
homework from bugwash: look at ministat and get more numbers